Python_backend/requirements_new.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# ML model dependencies
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
openai-whisper==20231117
einops==0.7.0

# Image and video processing
pillow==10.1.0
opencv-python==4.8.1.78

# Data validation and serialization
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0
psutil==5.9.6
numpy==1.24.3

# Optional: For better performance
# accelerate==0.24.1  # Uncomment if you want model optimization
